0830      splash activity 완성      login activity 비밀번호하고 자동 로그인 남았다 시댕      register activity 완성      메인 activity xml 완료      메인 activity 자바 짜는거 포기하고 싶다

+++ app/src/main/java/com/example/edcan3vacationproject/GPSdata.java
@@ -1,10 +1,10 @@
 package com.example.edcan3vacationproject;
 
 public class GPSdata  {
-    public float Longitude;
-    public float Latitude;
+    public double Longitude;
+    public double Latitude;
 
-    public GPSdata(float longitude, float latitude)
+    public GPSdata(double longitude, double latitude)
     {
         Longitude = longitude;
         Latitude = latitude;
